Files reorganization
